Moved versioning to a separate module, and avoid import of the private proto module
