+++ src/python_pachyderm/__init__.py
@@ -1,5 +1,2 @@
-# -*- coding: utf-8 -*-
-
 from .pfs import PfsClient
 from .pps import PpsClient
-from .util import get_remote_version
