+++ src/python_pachyderm/util.py
@@ -1,12 +1,6 @@
 import os
 
 from python_pachyderm._proto.pfs import pfs_pb2 as pfs_proto
-from python_pachyderm._proto.version.versionpb.version_pb2_grpc import (
-    google_dot_protobuf_dot_empty__pb2 as pb_empty,
-    APIStub as VersionStub,
-    grpc
-)
-from contextlib import closing
 
 
 def get_address(host=None, port=None):
@@ -38,10 +32,3 @@ def commit_from(src, allow_just_repo=False):
     if not allow_just_repo:
         raise ValueError("Invalid commit type")
     return pfs_proto.Commit(repo=pfs_proto.Repo(name=src))
-
-
-def get_remote_version(host=None, port=None):
-    with closing(grpc.insecure_channel(get_address(host, port))) as channel:
-        stub = VersionStub(channel)
-        version = stub.GetVersion(pb_empty.Empty())
-        return version
