+++ src/python_pachyderm/version/__init__.py
@@ -1,0 +1,2 @@
+from .client import get_remote_version
+from python_pachyderm._proto.version.versionpb.version_pb2 import *
