+++ tests/test_get_version.py
@@ -1,14 +1,13 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
 
 """Tests for the `get_remote_version` function of the `python_pachyderm` package."""
-import python_pachyderm
-from python_pachyderm._proto.version.versionpb.version_pb2 import Version
+
+from python_pachyderm import version
 
 
 def test_pfs_client_init_with_default_host_port():
     # GIVEN a Pachyderm deployment
     # WHEN a client requests the remote version of a pachyderm deployment
-    version_pb = python_pachyderm.get_remote_version()
+    version_pb = version.get_remote_version()
     # THEN remote version should be a version object
-    assert isinstance(version_pb, Version)
+    assert isinstance(version_pb, version.Version)
