+++ tests/test_pps_client.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
 
 """Tests for the `PpsClient` class of the `python_pachyderm` package."""
 
@@ -9,17 +8,15 @@ import pytest
 import random
 import string
 
-import python_pachyderm
-from python_pachyderm._proto.pfs import pfs_pb2 as pfs_proto
-from python_pachyderm._proto.pps import pps_pb2 as pps_proto
+from python_pachyderm import pfs, pps
 
 def random_string(n):
     return "".join(random.choice(string.ascii_lowercase + string.digits) for _ in range(n))
 
 class Sandbox:
     def __init__(self, test_name):
-        pfs_client = python_pachyderm.PfsClient()
-        pps_client = python_pachyderm.PpsClient()
+        pfs_client = pfs.PfsClient()
+        pps_client = pps.PpsClient()
 
         repo_name_suffix = random_string(6)
         input_repo_name = "{}-input-{}".format(test_name, repo_name_suffix)
@@ -29,8 +26,8 @@ class Sandbox:
 
         pps_client.create_pipeline(
             pipeline_repo_name,
-            transform=pps_proto.Transform(cmd=["sh"], image="alpine", stdin=["cp /pfs/{}/*.dat /pfs/out/".format(input_repo_name)]),
-            input=pps_proto.Input(pfs=pps_proto.PFSInput(glob="/*", repo=input_repo_name)),
+            transform=pps.Transform(cmd=["sh"], image="alpine", stdin=["cp /pfs/{}/*.dat /pfs/out/".format(input_repo_name)]),
+            input=pps.Input(pfs=pps.PFSInput(glob="/*", repo=input_repo_name)),
             enable_stats=True,
         )
 
@@ -45,7 +42,7 @@ class Sandbox:
 
     def wait_for_job(self):
         # block until the commit is ready
-        self.pfs_client.inspect_commit(self.commit, block_state=pfs_proto.READY)
+        self.pfs_client.inspect_commit(self.commit, block_state=pfs.READY)
 
         # while the commit is ready, the job might not be listed on the first
         # call, so repeatedly list jobs until it's available
@@ -94,7 +91,7 @@ def test_stop_job():
     except grpc._channel._Rendezvous:
         # if it failed, it should be because the job already finished
         job = sandbox.pps_client.inspect_job(job_id)
-        assert job.state == pps_proto.JOB_SUCCESS
+        assert job.state == pps.JOB_SUCCESS
     else:
         # This is necessary because `StopJob` does not wait for the job to be
         # killed before returning a result.
@@ -102,7 +99,7 @@ def test_stop_job():
         # https://github.com/pachyderm/pachyderm/issues/3856
         time.sleep(1)
         job = sandbox.pps_client.inspect_job(job_id)
-        assert job.state == pps_proto.JOB_KILLED
+        assert job.state == pps.JOB_KILLED
 
 def test_delete_job():
     sandbox = Sandbox("delete_job")
@@ -121,7 +118,7 @@ def test_datums():
     datums = list(sandbox.pps_client.list_datum(job_id))
     assert len(datums) == 1
     datum = sandbox.pps_client.inspect_datum(job_id, datums[0].datum_info.datum.id)
-    assert datum.state == pps_proto.SUCCESS
+    assert datum.state == pps.SUCCESS
 
     # Just ensure this doesn't raise an exception
     sandbox.pps_client.restart_datum(job_id)
